updating itoa

+++ ft_itoa.c
@@ -6,7 +6,7 @@
 /*   By: apaduan- <apaduan-@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/12 21:40:58 by apaduan-          #+#    #+#             */
-/*   Updated: 2021/06/17 23:43:49 by apaduan-         ###   ########.fr       */
+/*   Updated: 2021/06/17 23:44:40 by apaduan-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@
 char	*ft_itoa(int n)
 {
 	unsigned int neg;
-	long int nbr;
+	long long int nbr;
 	int digits;
 	char *dest;
